[IMP] Refactor and remove utils.py
Create dedicated class for web services
Create dedicated class for xml_utils (static helper methods)
Quick TODO review
