+++ addons/l10n_es_edi_tbai/demo/demo_company.xml
@@ -3,7 +3,7 @@
     <data>
         <record id="l10n_es.demo_company_es" model="res.company">
             <field name="l10n_es_tbai_test_env" eval="True"/>
-            <field name="vat" model="l10n_es.edi.tbai.util" eval="obj()._random_vat()"/>
+            <field name="vat" model="l10n_es.edi.tbai.misc_util" eval="obj()._random_vat()"/>
         </record>
         <record id="base.main_company" model="res.company">
             <field name="l10n_es_tbai_test_env" eval="True"/>
